FIRST STEP - Compiled Version - Updated all SRC
It's a first step, so it's just a simple scene with a basic reproduction of the sun, done with a generic texture taken online applied on a pbr emissive material.
Made it rotate using a sawtooth wave and pulse a little using a LFO, applied also to the bloom post fx.
I'm not sure how to handle licenses on the src code, so I'll refer to all code's owner licenses!
Commercial use is not allowed without a license.